+++ src/main/java/com/example/dernierespoirsae/Observateur/ObservateurPositionY.java
@@ -20,6 +20,6 @@ public class ObservateurPositionY implements ChangeListener<Number> {
     public void changed(ObservableValue<? extends Number> obs, Number oldV, Number newV) {
 
         if(this.joueur.getY() > Main.largeur/2 && this.joueur.getY() < this.joueur.getEnvironnement().getInfoTuile()[0] * this.joueur.getEnvironnement().getInfoTuile()[2] - Main.largeur/2) //300 = la moitié de la largeur de la Scene
-            this.principalPane.setTranslateY(Main.largeur/2-joueur.getY());
+            this.principalPane.setTranslateY((double) Main.largeur /2-joueur.getY());
     }
 }
